Made some better histogram plots of the 30 mn temp para treatments. Made a column that combined temperature treatments and emergence status, used that to make a count histogram that plots lines instead of bars (geom_freqpoly).
Made some plots of mean wasp survival to emergence and eclosion, where mean temp is on the x axis and temp var is the grouping variable--will use this for mss instead of the other way around. Made mean temp numeric instead of factor, so it reflects the accurate scaling between the temperatures. Need to do this for development time as well.